Update data-lifecycle.md
Here is draft text for the Data Life Cycle page. Please feel free to improve on this, and also correct anything that is wrong or could be worded better.  Add more details too if you want!  I committed it to master, so just edit it directly and re-commit to  master or make a new branch - whatever you prefer.